Rename NotesModal to CommentsModal
